@@ -12,10 +12,10 @@
     ],
     "license": "BSD-2-Clause",
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=7.2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.2.*"
+        "phpunit/phpunit": "^7"
     },
     "autoload": {
         "psr-0" : {
