@@ -730,6 +730,16 @@ class Metas implements HasFilters, HasActions {
 				'level'      => array( 'feed' ),
 				'serialize'  => false,
 			),
+			'pf_feed_default_author'  => array(
+				'name'       => 'pf_feed_default_author',
+				'title'      => __( 'Default Feed Author', 'pf' ),
+				'definition' => __( 'The default author set for items in the feed', 'pf' ),
+				'function'   => __( 'Stores the default author that is used when no author is available in the feed.', 'pf' ),
+				'type'       => array( 'adm' ),
+				'use'        => array( 'api' ),
+				'level'      => array( 'feed' ),
+				'serialize'  => false,
+			),
 			'feedUrl'                 => array(
 				'name'       => 'feedUrl',
 				'title'      => __( 'Feed URL', 'pf' ),
@@ -825,6 +835,61 @@ class Metas implements HasFilters, HasActions {
 		$metas = apply_filters( 'pf_meta_terms', $metas );
 
 		return $metas;
+	}
+
+	public function map_post_type_to_level($post_type){
+		$mapping = array(
+				'feed'       => 'pf_feed',
+				'item'      => 'pf_feed_item',
+				'nomination' => 'nomination',
+				'post'   => 'post'
+		);
+		$mapping = array(
+				'pf_feed'       => 'feed',
+				'pf_feed_item'      => 'item',
+				'nomination' => 'nomination',
+				'post'   => 'post'
+		);
+		$mapping = apply_filters( 'pf_post_type_to_level', $mapping );
+		if (array_key_exists($post_type, $mapping)){
+			return $mapping[$post_type];
+		} else {
+			return false;
+		}
+	}
+
+	public function attach_metas_by_use($post_object, $use = 'api', $admin = false) {
+		$post_id = $post_object->ID;
+		foreach ( $this->structure() as $key => $meta ) {
+			// var_dump($key);
+			if (in_array($use, $meta['use']) ){
+				if (in_array($this->map_post_type_to_level($post_object->post_type), $meta['level'])) {
+					if (false === $admin){
+						if (!in_array('adm', $meta['type']) ){
+							if (!property_exists($post_object, $key)){
+								// var_dump('yes'); die();
+								$post_object->$key = $this->get_post_pf_meta($post_id, $meta['name']);
+							} else {
+								// var_dump('set fail');
+							}
+						} else {
+							// var_dump('adm fail');
+						}
+					} else {
+						// var_dump('oops'); die();
+							if (!isset($post_object->$key)){
+								$post_object->$key = $this->get_post_pf_meta($post_id, $meta['name']);
+							}
+					}
+				} else {
+					// var_dump($this->map_post_type_to_level($post_object->post_type));
+				}
+			} else {
+				// var_dump( 'no' );
+			}
+		}
+		// die();
+		return $post_object;
 	}
 
 	/**
@@ -949,6 +1014,9 @@ class Metas implements HasFilters, HasActions {
 		$meta = $this->meta_interface->get_meta( $id, $field, $single );
 		if ( $serialized ) {
 			if ( empty( $meta ) || ! array_key_exists( $key, $meta ) ) {
+				if ( !is_array( $meta ) ) {
+					$meta = [];
+				}
 				$old_meta     = $this->meta_interface->get_meta( $id, $key, $single );
 				$meta[ $key ] = $old_meta;
 				$this->meta_interface->update_meta( $id, $field, $meta );
@@ -988,6 +1056,16 @@ class Metas implements HasFilters, HasActions {
 				if ( empty( $meta_value ) || is_wp_error( $meta_value ) ) {
 					$meta_value = get_the_source_title( $id );
 				}
+				break;
+
+			case 'item_author':
+				if ( empty( $meta_value ) || 'aggregation' === $meta_value ) {
+					$parent_value = pressforward( 'controller.metas' )->get_post_pf_meta( $id, 'pf_feed_default_author', true );
+					if ( ! empty( $parent_value ) ) {
+						$meta_value = $parent_value;
+					}
+				}
+				break;
 
 				// no break
 			default:
@@ -1158,9 +1236,13 @@ class Metas implements HasFilters, HasActions {
 		return $check;
 	}
 
-	public function forward_to_origin_status( $ID, $check = true, $the_value = false ) {
+	public function forward_to_origin_status( $id, $check = true, $the_value = false ) {
+		$item_id = pressforward( 'controller.metas' )->get_post_pf_meta( $id, 'item_id', true );
+		if ( empty( $item_id ) ) {
+			return 'no-forward';
+		}
 		if ( $check ) {
-			$value = pressforward( 'controller.metas' )->get_post_pf_meta( $ID, 'pf_forward_to_origin', true );
+			$value = pressforward( 'controller.metas' )->get_post_pf_meta( $id, 'pf_forward_to_origin', true );
 		} else {
 			$value = $the_value;
 		}
@@ -1183,6 +1265,10 @@ class Metas implements HasFilters, HasActions {
 		remove_filter( 'get_post_metadata', array( $this, 'usable_forward_to_origin_status' ), 10 );
 		$value = $this->forward_to_origin_status( $object_id );
 		add_filter( 'get_post_metadata', array( $this, 'usable_forward_to_origin_status' ), 10, 4 );
+
+		if ( ! $single ) {
+			$value = array( $value );
+		}
 
 		return $value;
 	}
