@@ -111,7 +111,7 @@ class Text
         $strText = trim(preg_replace('`[ ]*([\.])`', '$1 ', $strText)); // Pad sentence terminators
 
         // Lower case all words following terminators (for gunning fog score)
-        $strText = preg_replace_callback('`\. [^\. ]`', create_function('$matches', 'return strtolower($matches[0]);'), $strText);
+        $strText = preg_replace_callback('`\. [^\. ]`', function($matches) { return strtolower($matches[0]); }, $strText);
 
         $strText = trim($strText);
 
@@ -260,7 +260,7 @@ class Text
         try {
 
             if (!self::$blnMbstring) {
-                throw new Exception('The extension mbstring is not loaded.');
+                throw new \Exception('The extension mbstring is not loaded.');
             }
 
             if ($strEncoding == '') {
@@ -288,7 +288,7 @@ class Text
         }
 
         // Will be tripped by em dashes with spaces either side, among other similar characters
-        $intWords = 1 + self::textLength(preg_replace('`[^ ]`', '', $strText), $strEncoding); // Space count + 1 is word count
+        $intWords = 1 + self::textLength(preg_replace('`[^ ]`', '', preg_replace('`\s+`', ' ', $strText)), $strEncoding); // Space count + 1 is word count
 
         return $intWords;
     }
