@@ -1,5 +1,5 @@
-jQuery(window).load(function() {
-	jQuery(".refreshfeed").click(function (evt){
+jQuery(window).load(function () {
+	jQuery(".refreshfeed").click(function (evt) {
 		evt.preventDefault();
 		jQuery('.loading-top').show();
 		jQuery('#responses').css('display', 'block');
@@ -7,32 +7,53 @@ jQuery(window).load(function() {
 		jQuery.post(ajaxurl, {
 			action: 'assemble_feed_for_pull'
 		},
-		function(response) {
-			//jQuery('.loading-top').hide();
-			//jQuery('#responses').css('display', 'block');
-			//console.log(response);
-			//response = JSON.parse(response);
-			console.log(response);
-			jQuery('#responses').html('<span class="basic-response">'+response.action_taken+'</span>');
-			jQuery('#errors').html(response);
-			//jQuery("#test-div1").append(data);
-		});
+			function (response) {
+				//jQuery('.loading-top').hide();
+				//jQuery('#responses').css('display', 'block');
+				//console.log(response);
+				//response = JSON.parse(response);
+				console.log(response);
+				jQuery('#responses').html('<span class="basic-response">' + response.action_taken + '</span>');
+				jQuery('#errors').html(response);
+				//jQuery("#test-div1").append(data);
+			});
 
 	});
 
-	jQuery(".cleanfeeds").click(function (evt){
+	jQuery(".cleanfeeds").click(function (evt) {
 		evt.preventDefault();
 		jQuery('.loading-top').show();
 		jQuery.post(ajaxurl, {
 			action: 'disassemble_item'
 		},
-		function(response) {
-			//jQuery('.loading-top').hide();
-			jQuery('#responses').css('display', 'block');
-			jQuery('#responses').html('<span class="basic-response">'+response.action_taken+'</span>');
-			jQuery('#errors').html(response);
-			//jQuery("#test-div1").append(data);
-		});
+			function (response) {
+				//jQuery('.loading-top').hide();
+				jQuery('#responses').css('display', 'block');
+				jQuery('#responses').html('<span class="basic-response">' + response.action_taken + '</span>');
+				jQuery('#errors').html(response);
+				//jQuery("#test-div1").append(data);
+			});
+
+	});
+
+	jQuery("#pressforward-nt__regenerate-button").click(function (evt) {
+		// evt.preventDefault();
+		// jQuery('.loading-top').show();
+		jQuery.post(ajaxurl, {
+			action: 'regenerate_user_keys'
+		},
+			function (response) {
+				//jQuery('.loading-top').hide();
+				console.log(jQuery('#pfnt__pfSiteData').val());
+				var settingsObject = JSON.parse(jQuery('#pfnt__pfSiteData').val());
+				// console.log('response', response);
+				// console.log('response object', response.response.response_data);
+				//var resultObject = JSON.parse(response.response_data);
+				settingsObject.ku = response.ku;
+				settingsObject.ki = response.ki;
+				jQuery('#pfnt__pfSiteData').val(JSON.stringify(settingsObject));
+				//jQuery("#test-div1").append(data);
+			});
 
 	});
 
