@@ -133,8 +133,11 @@ class PF_Debugger extends PF_Module {
 		$ntp_args = array( 'posts_per_page' => -1, 'meta_key' => 'item_link' );
 
 		$nominated_to_posts = get_posts( $ntp_args );
-		$nomed_posts = count( $nominated_to_posts );
-
+		if ( is_array( $nominated_to_posts ) ){
+			$nomed_posts = count( $nominated_to_posts );
+		} else {
+			$nomed_posts = 0;
+		}
 		?>
 		<div class="wrap">
 			<h2>Current Log</h2>
